changes following code review.

+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/BaseHapiFhirResourceDao.java
@@ -417,7 +417,7 @@ public abstract class BaseHapiFhirResourceDao<T extends IBaseResource> extends B
 
 		// Pre-cache the match URL
 		if (theMatchUrl != null) {
-			myResourceSearchUrlSvc.storeMatchUrlForResource(getResourceName(), theMatchUrl, jpaPid);
+			myResourceSearchUrlSvc.enforceMatchUrlResourceUniqueness(getResourceName(), theMatchUrl, jpaPid);
 			myMatchResourceUrlService.matchUrlResolved(theTransactionDetails, getResourceName(), theMatchUrl, jpaPid);
 		}
 
@@ -732,10 +732,10 @@ public abstract class BaseHapiFhirResourceDao<T extends IBaseResource> extends B
 		return retVal;
 	}
 
-	@Override
 	protected ResourceTable updateEntityForDelete(RequestDetails theRequest, TransactionDetails theTransactionDetails, ResourceTable theEntity) {
 		myResourceSearchUrlSvc.deleteByResId(theEntity.getId());
-		return super.updateEntityForDelete(theRequest, theTransactionDetails, theEntity);
+		Date updateTime = new Date();
+		return updateEntity(theRequest, null, theEntity, updateTime, true, true, theTransactionDetails, false, true);
 	}
 
 	private void validateDeleteEnabled() {
@@ -1767,7 +1767,13 @@ public abstract class BaseHapiFhirResourceDao<T extends IBaseResource> extends B
 
 	@Override
 	protected DaoMethodOutcome doUpdateForUpdateOrPatch(RequestDetails theRequest, IIdType theResourceId, String theMatchUrl, boolean thePerformIndexing, boolean theForceUpdateVersion, T theResource, IBasePersistedResource theEntity, RestOperationTypeEnum theOperationType, TransactionDetails theTransactionDetails) {
-		myResourceSearchUrlSvc.deleteByResId((Long) theEntity.getPersistentId().getId());
+
+		// we stored a resource searchUrl at creation time to prevent resource duplication.  Let's remove the entry on the
+		// first update but guard against unnecessary trips to the database on subsequent ones.
+		if(theEntity.getVersion() < 2){
+			myResourceSearchUrlSvc.deleteByResId((Long) theEntity.getPersistentId().getId());
+		}
+
 		return super.doUpdateForUpdateOrPatch(theRequest, theResourceId, theMatchUrl, thePerformIndexing, theForceUpdateVersion, theResource, theEntity, theOperationType, theTransactionDetails);
 	}
